Overhaul, added an offset option
This is a major overhaul. Now there are two options (A and B). A is the original; the new one (B) has the bolt offset and a hole for a retaining screw. No glue required. Option A or B can be chosen via the customizer.
Also fixed customizer parameter comments and hid non-customizable variables.